Server/requirements.txt
absl-py==2.2.2
astunparse==1.6.3
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cycler==0.12.1
Flask==3.1.0
flask-cors==5.0.1
flatbuffers==25.2.10
fonttools==4.49.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
keras==3.0.0
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.8.3
mdurl==0.1.2
ml_dtypes==0.5.1
namex==0.0.8
numpy==2.0.2
opt_einsum==3.4.0
optree==0.15.0
packaging==24.2
paho-mqtt==2.1.0
pillow==11.2.1
protobuf==5.29.4
Pygments==2.19.1
pyparsing==3.1.2
python-dateutil==2.8.2
python-dotenv==1.1.0
requests==2.32.3
rich==14.0.0
six==1.17.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.0.1
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.21.0
